Implemented Collision
It took a while to get collision to work, as I didn't realise I was
drawing Link's placeholder at its WORLD position, rather than zone
position. As a result, he was moving either 16 or 11 times faster than
he appeared to be, depending on his axis. This meant that he seemed to
be colliding with invisible walls.

Once I realised this, I modified the draw code to, instead of drawing
link's pixel position, drawing his tile position, while keeping him in
the correct spot on-screen.

Now I had the movement on screen actually accurate to in the stored
positions, I was able to notice that since collision was centred on the
top-left point of the sprite, Link could walk almost entirely inside any
bottom or right walls of rooms/corridors.

To fix this, I moved the start point of collision checking to the centre
of the sprite and using an offset of 8 (half the sprite) so any movement
started from the edge nearest the direction of movement.