Zone Transitions
While working on Zone transitions I discovered the transitions are
actually 15 and 10, not 16 and 11. The player doesn't move during them.

While watching a longplay to discover this, I also realised only the
bottom half of Link collides, the top half is allowed to clip as, since
Link is drawn after the world, he just appears to be in front of the
objects instead of clipping behind them